+++ src/demo/App.js
@@ -45,6 +45,7 @@ export default function App() {
                     options={{
                         rowInfoPosition: "right",
                         rowLinkPosition: "left",
+                        colToolsPosition: "bottom",
                         infoAttrPrimary: 5,
                         infoAttrSecondary: 6
                     }}